- Commit inicial

+++ jurosSimples.py
@@ -2,7 +2,7 @@
 # Apresentação
 print('\n\t\t\t  -- Calculadora de Juros Simples --')
 # Entradas
-c = int(input('Informe o capital (R$): '))
+c = float(input('Informe o capital (R$): '))
 i = float(input('Informe a taxa (%): '))
 n = int(input('Informe a quantidade de parcelas: '))
 # Processamento
